requirements.txt
requests==2.31.0
google-cloud-storage==1.44.0
pandas==2.2.3
fsspec==2025.2.0
gcsfs==2025.2.0
# google-cloud-bigquery==3.30.0
# google-cloud==0.34.0
# google-api-python-client==2.51.0